requirements.txt
3to2==1.1.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.7.0
attrs==26.1.0
beautifulsoup4==4.9.3
blinker==1.9.0
Brlapi==0.8.2
brotli==1.2.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.6
chrome-gnome-shell==0.0.0
click==8.1.8
configargparse==1.7.5
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
distro-info==1.0+deb11u1
exceptiongroup==1.3.1
flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.9
gherkin-official==29.0.0
greenlet==3.0.3
html5lib==1.1
idna==2.10
importlib-metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
locust==2.34.0
louis==3.16.0
lxml==4.6.3
Mako==1.1.3
markupsafe==3.0.3
msgpack==1.1.2
numpy==1.19.5
packaging==26.0
parse==1.21.1
parse-type==0.6.6
playwright==1.42.0
pluggy==1.6.0
psutil==7.2.2
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
pydantic==2.6.4
pydantic-core==2.16.3
pyee==11.0.1
pygments==2.19.2
PyGObject==3.38.0
pysmbc==1.0.23
pytest==7.4.4
pytest-base-url==2.1.0
pytest-bdd==7.1.2
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.4.4
python-apt==2.2.1.1
python-debian==0.1.39
python-dotenv==1.0.1
python-slugify==8.0.4
python-xapp==2.0.2
python-xlib==0.29
pyxdg==0.27
pyzmq==27.1.0
requests==2.31.0
scour==0.38.2
setproctitle==1.2.1
six==1.16.0
soupsieve==2.2.1
text-unidecode==1.3
tinycss==0.4
tomli==2.4.0
typing-extensions==4.15.0
unattended-upgrades==0.1
urllib3==1.26.5
webencodings==0.5.1
werkzeug==3.1.7
xdg==5
zipp==3.23.0
zope-event==6.0
zope-interface==8.0.1
